improvement: use pc for finding references of local symbols
