+++ mtags/src/main/scala-2/scala/meta/internal/pc/PcCollector.scala
@@ -5,291 +5,78 @@ import scala.reflect.internal.util.RangePosition
 import scala.meta.pc.OffsetParams
 import scala.meta.pc.VirtualFileParams
 
-abstract class PcCollector[T](
-    val compiler: MetalsGlobal,
-    params: VirtualFileParams
-) {
+trait PcCollector[T] { self: WithCompilationUnit =>
   import compiler._
 
   def collect(
       parent: Option[Tree]
   )(tree: Tree, pos: Position, sym: Option[Symbol]): T
-  val unit: RichCompilationUnit = addCompilationUnit(
-    code = params.text(),
-    filename = params.uri().toString(),
-    cursor = None
-  )
-  val offset: Int = params match {
-    case p: OffsetParams => p.offset()
-    case _: VirtualFileParams => 0
-  }
-  val pos: Position = unit.position(offset)
-  lazy val text = unit.source.content
-  private val caseClassSynthetics: Set[Name] = Set(nme.apply, nme.copy)
-  typeCheck(unit)
-  lazy val typedTree: Tree = locateTree(pos) match {
-    // Check actual object if apply is synthetic
-    case sel @ Select(qual, name) if name == nme.apply && qual.pos == sel.pos =>
-      qual
-    case Import(expr, _) if expr.pos.includes(pos) =>
-      // imports seem to be marked as transparent
-      locateTree(pos, expr, acceptTransparent = true)
-    case t => t
-  }
 
-  /**
-   * Find all symbols that should be shown together.
-   * For example class we want to also show companion object.
-   *
-   * @param sym symbol to find the alternative candidates for
-   * @return set of possible symbols
-   */
-  def symbolAlternatives(sym: Symbol): Set[Symbol] = {
-    val all =
-      if (sym.isClass) {
-        if (sym.owner.isMethod) Set(sym) ++ sym.localCompanion(pos)
-        else Set(sym, sym.companionModule, sym.companion.moduleClass)
-      } else if (sym.isModuleOrModuleClass) {
-        if (sym.owner.isMethod) Set(sym) ++ sym.localCompanion(pos)
-        else Set(sym, sym.companionClass, sym.moduleClass)
-      } else if (sym.isTerm && (sym.owner.isClass || sym.owner.isConstructor)) {
-        val info =
-          if (sym.owner.isClass) sym.owner.info
-          else sym.owner.owner.info
-        Set(
-          sym,
-          info.member(sym.getterName),
-          info.member(sym.setterName),
-          info.member(sym.localName)
-        ) ++ constructorParam(sym) ++ sym.allOverriddenSymbols.toSet
-      } else Set(sym)
-    all.filter(s => s != NoSymbol && !s.isError)
-  }
-
-  private def constructorParam(
-      symbol: Symbol
-  ): Set[Symbol] = {
-    if (symbol.owner.isClass) {
-      val info = symbol.owner.info.member(nme.CONSTRUCTOR).info
-      info.paramss.flatten.find(_.name == symbol.name).toSet
-    } else Set.empty
-  }
-
-  private lazy val namedArgCache = {
-    val parsedTree = parseTree(unit.source)
-    parsedTree.collect { case arg @ AssignOrNamedArg(_, rhs) =>
-      rhs.pos -> arg
-    }.toMap
-  }
-  def fallbackSymbol(name: Name, pos: Position): Option[Symbol] = {
-    val context = doLocateImportContext(pos)
-    context.lookupSymbol(name, sym => sym.isType) match {
-      case LookupSucceeded(_, symbol) =>
-        Some(symbol)
-      case _ => None
-    }
-  }
-
-  // First identify the symbol we are at, comments identify @@ as current cursor position
-  lazy val soughtSymbols: Option[(Set[Symbol], Position)] = typedTree match {
-    /* simple identifier:
-     * val a = val@@ue + value
+  def resultWithSought(sought: Set[Symbol]): List[T] = {
+    val owners = sought
+      .map(_.owner)
+      .flatMap(o => symbolAlternatives(o))
+      .filter(_ != NoSymbol)
+    val soughtNames: Set[Name] = sought.map(_.name)
+    /*
+     * For comprehensions have two owners, one for the enumerators and one for
+     * yield. This is a heuristic to find that out.
      */
-    case (id: Ident) =>
-      // might happen in type trees
-      // also this doesn't seem to be picked up by semanticdb
-      if (id.symbol == NoSymbol)
-        fallbackSymbol(id.name, pos).map(sym =>
-          (symbolAlternatives(sym), id.pos)
+    def isForComprehensionOwner(named: NameTree) = {
+      if (named.symbol.pos.isDefined) {
+        def alternativeSymbol = sought.exists(symbol =>
+          symbol.name == named.name &&
+            symbol.pos.isDefined &&
+            symbol.pos.start == named.symbol.pos.start
         )
-      else {
-        Some(symbolAlternatives(id.symbol), id.pos)
-      }
-    /* Anonynous function parameters such as:
-     * List(1).map{ <<abc>>: Int => abc}
-     * In this case, parameter has incorrect namePosition, so we need to handle it separately
-     */
-    case (vd: ValDef) if isAnonFunctionParam(vd) =>
-      val namePos = vd.pos.withEnd(vd.pos.start + vd.name.length)
-      if (namePos.includes(pos)) Some(symbolAlternatives(vd.symbol), namePos)
-      else None
-
-    /* all definitions:
-     * def fo@@o = ???
-     * class Fo@@o = ???
-     * etc.
-     */
-    case (df: DefTree) if df.namePosition.includes(pos) =>
-      Some(symbolAlternatives(df.symbol), df.namePosition)
-    /* Import selectors:
-     * import scala.util.Tr@@y
-     */
-    case (imp: Import) if imp.pos.includes(pos) =>
-      imp.selector(pos).map(sym => (symbolAlternatives(sym), sym.pos))
-    /* simple selector:
-     * object.val@@ue
-     */
-    case (sel: NameTree) if sel.namePosition.includes(pos) =>
-      Some(symbolAlternatives(sel.symbol), sel.namePosition)
-
-    // needed for classOf[AB@@C]`
-    case lit @ Literal(Constant(TypeRef(_, sym, _))) if lit.pos.includes(pos) =>
-      val posStart = text.indexOfSlice(sym.decodedName, lit.pos.start)
-      if (posStart == -1) None
-      else
-        Some(
-          (
-            symbolAlternatives(sym),
-            new RangePosition(
-              lit.pos.source,
-              posStart,
-              lit.pos.point,
-              posStart + sym.decodedName.length
-            )
+        def sameOwner = {
+          val owner = named.symbol.owner
+          owner.isAnonymousFunction && owners.exists(o =>
+            pos.isDefined && o.pos.isDefined && o.pos.point == owner.pos.point
           )
-        )
-    /* named argument, which is a bit complex:
-     * foo(nam@@e = "123")
-     */
-    case _ =>
-      val apply = typedTree match {
-        case (apply: Apply) => Some(apply)
-        /**
-         * For methods with multiple parameter lists and default args, the tree looks like this:
-         * Block(List(val x&1, val x&2, ...), Apply(<<method>>, List(x&1, x&2, ...)))
-         */
-        case _ =>
-          typedTree.children.collectFirst {
-            case Block(_, apply: Apply) if apply.pos.includes(pos) =>
-              apply
-          }
+        }
+        alternativeSymbol && sameOwner
+      } else false
+    }
+
+    def soughtOrOverride(sym: Symbol) =
+      sought(sym) || sym.allOverriddenSymbols.exists(sought(_))
+
+    def soughtTreeFilter(tree: Tree): Boolean =
+      tree match {
+        case ident: Ident
+            if (soughtOrOverride(ident.symbol) ||
+              isForComprehensionOwner(ident)) =>
+          true
+        case tpe: TypeTree =>
+          sought(tpe.original.symbol)
+        case sel: Select =>
+          soughtOrOverride(sel.symbol)
+        case df: MemberDef =>
+          (soughtOrOverride(df.symbol) ||
+          isForComprehensionOwner(df))
+        case appl: Apply =>
+          appl.symbol != null &&
+          (owners(appl.symbol) ||
+            symbolAlternatives(appl.symbol.owner).exists(owners(_)))
+        case imp: Import =>
+          owners(imp.expr.symbol) && imp.selectors
+            .exists(sel => soughtNames(sel.name))
+        case bind: Bind =>
+          (soughtOrOverride(bind.symbol)) ||
+          isForComprehensionOwner(bind)
+        case _ => false
       }
-      apply
-        .collect {
-          case apply if apply.symbol != null =>
-            collectArguments(apply)
-              .flatMap(arg => namedArgCache.find(_._1.includes(arg.pos)))
-              .collectFirst {
-                case (_, AssignOrNamedArg(id: Ident, _))
-                    if id.pos.includes(pos) =>
-                  apply.symbol.paramss.flatten.find(_.name == id.name).map {
-                    s =>
-                      // if it's a case class we need to look for parameters also
-                      if (
-                        caseClassSynthetics(s.owner.name) && s.owner.isSynthetic
-                      ) {
-                        val applyOwner = s.owner.owner
-                        val constructorOwner =
-                          if (applyOwner.isCaseClass) applyOwner
-                          else
-                            applyOwner.companion match {
-                              case NoSymbol =>
-                                applyOwner
-                                  .localCompanion(pos)
-                                  .getOrElse(NoSymbol)
-                              case comp => comp
-                            }
-                        val info = constructorOwner.info
-                        val constructorParams = info.members
-                          .filter(_.isConstructor)
-                          .flatMap(_.paramss)
-                          .flatten
-                          .filter(_.name == id.name)
-                          .toSet
-                        (
-                          (constructorParams ++ Set(
-                            s,
-                            info.member(s.getterName),
-                            info.member(s.setterName),
-                            info.member(s.localName)
-                          )).filter(_ != NoSymbol),
-                          id.pos
-                        )
-                      } else (Set(s), id.pos)
-                  }
-              }
-              .flatten
-        }
-        .getOrElse(None)
-  }
 
-  def result(): List[T] = {
-    params match {
-      case _: OffsetParams => resultWithSought()
-      case _ => resultAllOccurences().toList
+    // 1. In most cases, we try to compare by Symbol#==.
+    // 2. If there is NoSymbol at a node, we check if the identifier there has the same decoded name.
+    //    It it does, we look up the symbol at this position using `fallbackSymbol` or `members`,
+    //    and again check if this time we got symbol equal by Symbol#==.
+    def soughtFilter(f: Symbol => Boolean): Boolean = {
+      sought.exists(f)
     }
-  }
 
-  def resultWithSought(): List[T] = {
-    soughtSymbols match {
-      case Some((sought, _)) =>
-        val owners = sought
-          .map(_.owner)
-          .flatMap(o => symbolAlternatives(o))
-          .filter(_ != NoSymbol)
-        val soughtNames: Set[Name] = sought.map(_.name)
-        /*
-         * For comprehensions have two owners, one for the enumerators and one for
-         * yield. This is a heuristic to find that out.
-         */
-        def isForComprehensionOwner(named: NameTree) = {
-          if (named.symbol.pos.isDefined) {
-            def alternativeSymbol = sought.exists(symbol =>
-              symbol.name == named.name &&
-                symbol.pos.isDefined &&
-                symbol.pos.start == named.symbol.pos.start
-            )
-            def sameOwner = {
-              val owner = named.symbol.owner
-              owner.isAnonymousFunction && owners.exists(o =>
-                pos.isDefined && o.pos.isDefined && o.pos.point == owner.pos.point
-              )
-            }
-            alternativeSymbol && sameOwner
-          } else false
-        }
-
-        def soughtOrOverride(sym: Symbol) =
-          sought(sym) || sym.allOverriddenSymbols.exists(sought(_))
-
-        def soughtTreeFilter(tree: Tree): Boolean =
-          tree match {
-            case ident: Ident
-                if (soughtOrOverride(ident.symbol) ||
-                  isForComprehensionOwner(ident)) =>
-              true
-            case tpe: TypeTree =>
-              sought(tpe.original.symbol)
-            case sel: Select =>
-              soughtOrOverride(sel.symbol)
-            case df: MemberDef =>
-              (soughtOrOverride(df.symbol) ||
-              isForComprehensionOwner(df))
-            case appl: Apply =>
-              appl.symbol != null &&
-              (owners(appl.symbol) ||
-                symbolAlternatives(appl.symbol.owner).exists(owners(_)))
-            case imp: Import =>
-              owners(imp.expr.symbol) && imp.selectors
-                .exists(sel => soughtNames(sel.name))
-            case bind: Bind =>
-              (soughtOrOverride(bind.symbol)) ||
-              isForComprehensionOwner(bind)
-            case _ => false
-          }
-
-        // 1. In most cases, we try to compare by Symbol#==.
-        // 2. If there is NoSymbol at a node, we check if the identifier there has the same decoded name.
-        //    It it does, we look up the symbol at this position using `fallbackSymbol` or `members`,
-        //    and again check if this time we got symbol equal by Symbol#==.
-        def soughtFilter(f: Symbol => Boolean): Boolean = {
-          sought.exists(f)
-        }
-
-        traverseSought(soughtTreeFilter, soughtFilter).toList
-
-      case None => Nil
-    }
+    traverseSought(soughtTreeFilter, soughtFilter).toList
   }
 
   def resultAllOccurences(): Set[T] = {
@@ -518,8 +305,7 @@ abstract class PcCollector[T](
           tree.children.foldLeft(acc)(traverse(_, _))
       }
     }
-    val all = traverseWithParent(None)(Set.empty[T], unit.lastBody)
-    all
+    traverseWithParent(None)(Set.empty[T], unit.lastBody)
   }
 
   private def annotationChildren(mdef: MemberDef): List[Tree] = {
@@ -540,14 +326,23 @@ abstract class PcCollector[T](
     }
   }
 
-  private def collectArguments(apply: Apply): List[Tree] = {
-    apply.fun match {
-      case appl: Apply => collectArguments(appl) ++ apply.args
-      case _ => apply.args
-    }
-  }
+}
 
-  private def isAnonFunctionParam(vd: ValDef): Boolean =
-    vd.symbol != null && vd.symbol.owner.isAnonymousFunction && vd.rhs.isEmpty
+abstract class SimpleCollector[T](
+    compiler: MetalsGlobal,
+    params: VirtualFileParams
+) extends WithCompilationUnit(compiler, params)
+    with PcCollector[T] {
+  def result(): List[T] = resultAllOccurences().toList
+}
 
+abstract class WithSymbolSearchCollector[T](
+    compiler: MetalsGlobal,
+    params: OffsetParams
+) extends WithCompilationUnit(compiler, params)
+    with PcCollector[T]
+    with PcSymbolSearch {
+  def result(): List[T] = soughtSymbols
+    .map { case (sought, _) => resultWithSought(sought) }
+    .getOrElse(Nil)
 }
