+++ mtags/src/main/scala-3/scala/meta/internal/pc/PcDocumentHighlightProvider.scala
@@ -13,7 +13,8 @@ import org.eclipse.lsp4j.DocumentHighlightKind
 final class PcDocumentHighlightProvider(
     driver: InteractiveDriver,
     params: OffsetParams,
-) extends PcCollector[DocumentHighlight](driver, params):
+) extends WithSymbolSearchCollector[DocumentHighlight](driver, params):
+
   def collect(
       parent: Option[Tree]
   )(
@@ -29,4 +30,5 @@ final class PcDocumentHighlightProvider(
 
   def highlights: List[DocumentHighlight] =
     result().distinctBy(_.getRange())
+
 end PcDocumentHighlightProvider
