+++ mtags/src/main/scala-3/scala/meta/internal/pc/PcSemanticTokensProvider.scala
@@ -58,7 +58,7 @@ final class PcSemanticTokensProvider(
         case _ => !df.rhs.isEmpty
     case _ => false
 
-  object Collector extends PcCollector[Option[Node]](driver, params):
+  object Collector extends SimpleCollector[Option[Node]](driver, params):
     override def collect(
         parent: Option[Tree]
     )(
