+++ tests/slow/src/test/scala/tests/maven/MavenLspSuite.scala
@@ -237,8 +237,9 @@ class MavenLspSuite extends BaseImportSuite("maven-import") {
         """.stripMargin,
       )
       // we should still have references despite fatal warning
+      refs <- server.workspaceReferences()
       _ = assertNoDiff(
-        server.workspaceReferences().references.map(_.symbol).mkString("\n"),
+        refs.references.map(_.symbol).sorted.mkString("\n"),
         """|_empty_/A.
            |_empty_/A.B.
            |_empty_/Warning.
